move checkForGameOver to snake helpers
